requirement.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.1.0
fonttools==4.55.8
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
logging==0.4.9.6
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
narwhals==1.25.1
nest-asyncio==1.6.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
stack-data==0.6.3
streamlit==1.43.2
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
